move into usual places
